Add ControllerBase + ViewBase
