+++ Index.html
@@ -97,7 +97,7 @@
                 </div>
                 <div class="advantages-about">
                     <div class="btn advantages-about-btn">
-                        <a href="#" class="about-btn">Узнать больше</a>
+                        <a href="#about" class="about-btn">Узнать больше</a>
                     </div>
                 </div>
                 <div class="travel-choice" id="choice">
