+++ scripts/Controllers/TestController.js
@@ -2,32 +2,14 @@
 /* jshint browser: true */
 /* jshint -W097 */
 
-export default class TestController{
+import ControllerBase from '/scripts/Controllers/ControllerBase.js';
+export default class TestController extends ControllerBase{
 
     constructor(view, model){
-        this._view = view;
-        this._model = model;
+        super(view,model);
         this._currentQuestion = 0;
         this._result = 0;
     }
-
-    get view(){
-        return this._view;
-    }
-
-    set view(value){
-        this._view = value;
-    }
-
-    get model(){
-        return this._model;
-    }
-
-    set model(value){
-        this._model = value;
-    }
-
-    //todo: get/set
 
     initialize(){
         this.bindEvent(this.nextQuestion);
